Made the contract of transform_func more concrete
This was a fairly minor overhaul but with fairly large repercussions.
The problem was that you can't @push(null) to a stream - that signifies
end-of-stream. So I was calling back with (null, new Buffer('')) when
I wanted to no-op inside of transform_func, but in an object mode
stream, instead of being a no-op it would write an empty buffer. Now, if
transform_func calls back with (null, null), it will simply skip the
@push call.